Add IsOverlayEnabled parameter to BootstrapSelect
Introduce `IsOverlayEnabled` parameter to control the display of an overlay when the dropdown is active. The overlay allows users to click outside the dropdown to close it, with a default value of `true`. Update CSS class from `.overlay` to `.bootstrap-select-overlay` and modify associated CSS and SCSS files for proper styling.